requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose[cryptography]
python-dotenv
alembic
qrcode
pillow
apscheduler
pytz
pytest
pytest-mock
